+SMTP config
You can pass environment variables from your shell straight through to a service’s containers with the ‘environment’ key by not giving them a value.
The value of the environment variable in the container is taken from the value for the same variable in the shell in which Compose is run.
Export variables in $HOME/.bashrc